# go/model_accomodation.go
/*
 * aid-ukraine
 *
 * Migration aid database and searcher for the Ukrain refugees.
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Accomodation struct {

	Id int64 `json:"id,omitempty"`

	Person *Person `json:"person,omitempty"`

	City string `json:"city,omitempty"`

	Max int64 `json:"max,omitempty"`

	Male bool `json:"male,omitempty"`

	Female bool `json:"female,omitempty"`

	Kids bool `json:"kids,omitempty"`

	Animals bool `json:"animals,omitempty"`

	AvailableFrom string `json:"availableFrom,omitempty"`

	Transport bool `json:"transport,omitempty"`

	Comment string `json:"comment,omitempty"`

	Status *Status `json:"status,omitempty"`

	StatusComment string `json:"statusComment,omitempty"`
}