# go/model_status.go
/*
 * aid-ukraine
 *
 * Migration aid database and searcher for the Ukrain refugees.
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Status struct {

	Id int64 `json:"id,omitempty"`

	Description string `json:"description,omitempty"`
}