# go/model_transportation.go
/*
 * aid-ukraine
 *
 * Migration aid database and searcher for the Ukrain refugees.
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Transportation struct {

	Id int64 `json:"id,omitempty"`

	Person *Person `json:"person,omitempty"`

	Departure string `json:"departure,omitempty"`

	Seats int64 `json:"seats,omitempty"`

	KidsSeat bool `json:"kidsSeat,omitempty"`

	Animals bool `json:"animals,omitempty"`

	AvailableFrom string `json:"availableFrom,omitempty"`

	Comment string `json:"comment,omitempty"`

	Status *Status `json:"status,omitempty"`

	StatusComment string `json:"statusComment,omitempty"`
}